requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==5.0.1
Django==4.2.7
django-allauth==0.58.1
django-braces==1.15.0
django-cors-headers==4.3.0
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.0
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
twilio==8.10.2
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
wrapt==1.15.0
yarl==1.9.3
